Fix local Docker Jekyll environment

+++ bin/entry_point.sh
@@ -1,21 +1,27 @@
 #!/bin/bash
+set -e
 
-CONFIG_FILE=_config.yml 
+echo "Entry point script running"
 
-/bin/bash -c "rm -f Gemfile.lock && exec jekyll serve --watch --port=8080 --host=0.0.0.0 --livereload --verbose --trace --force_polling"&
+CONFIG_FILE=_config.yml
+JEKYLL_CMD="bundle exec jekyll serve --watch --port=8080 --host=0.0.0.0 --livereload --verbose --trace --force_polling"
+
+/bin/bash -c "exec $JEKYLL_CMD"&
 
 while true; do
 
   inotifywait -q -e modify,move,create,delete $CONFIG_FILE
 
   if [ $? -eq 0 ]; then
- 
+
     echo "Change detected to $CONFIG_FILE, restarting Jekyll"
 
     jekyll_pid=$(pgrep -f jekyll)
-    kill -KILL $jekyll_pid
+    if [ -n "$jekyll_pid" ]; then
+      kill -KILL $jekyll_pid
+    fi
 
-    /bin/bash -c "rm -f Gemfile.lock && exec jekyll serve --watch --port=8080 --host=0.0.0.0 --livereload --verbose --trace --force_polling"&
+    /bin/bash -c "exec $JEKYLL_CMD"&
 
   fi
